Removed the flag argument
The flag argument was used to differ the behavior of the program between getting a directory or getting a user, but there is no need to do that.
If the request is for a user directory i.e. /home/<user>/sws if the sws directory has an index.html, serve it.